Proof object generation (WIP) (#230)
In this (rather large) PR, we begin quite a serious reworking of the proof object extraction mechanism.
We introduce a nominal/named version of matching logic patterns, which uses names for variables.
We also have a named version for proofs. We provide a nontrivial translation function from the locally nameless representation of patterns into named representation of patterns; this translation function should preserve proofs.
We begin a work on showing that it actually does so. For this purpose, we prove several lemmas about the pattern translation function.

Co-authored-by: Lucas Pena <lucaspena13@gmail.com>