remove wflexprod.v and some old code (#334)
This removes some code we wanted to use for a tautology solver some time ago (maybe two years?) - but we then discovered that this approach does not work for us. The code remained because we thought we could reuse it. But the proofs are so ugly that I no longer want to maintain them.

* remove wflexprod.v

* remove old, unused code