It might work
I found a way to prove well-foundedness for rose trees and it works,
but the same strategy fails for my patterns.

But I found that if I define a dependent rose tree, *similar* to
the application pattern, it does work. So I'm starting over in a clean
file, making it incrementally more complex to see *where* it fails.
Also the original file is way too crowded.

I'm also using the Arguments command liberally to hopefully make it
easier to deal with the dependent stuff. It's not harming me so far.

TODO: replace OpmlPattern.v with doover.v and delete RT.v when done