Progress with subsorts
First I tried a version of the carrier that mirrors the subsort
relation, but it didn't seem to be very helpful and didn't solve the
problem of not being able to do inversion on Prop.

So I decided to see what would happen if we used a computable relation
(crelation) for subsorting, which did allow me to do inversion and
define the model. For this, I used the original carrier as that seems
easier to generate automatically.

I tried some axioms, and one was provable but the others are really
ugly. I'm not sure I even defined them correctly (see ? comments) but I
suspect they won't simplify because of the ltac definition. However, due
to the excessive use of dependent destructions/inductions I'm not sure
if a gallina definition would look much better.

I also tried an EqDecision definition, should be possible, but
transports don't make it easy, and I didn't want to spend more time on
it. Only needed for list inclusion.