Closer to understanding problem
The second axiom simplifies down to app without problems, but since one
of its parameters is an evar_valuation, all the dependent inductions
don't simplify and it creates a mess.

With the non-meta list approach this didn't come up because it was
enough to know that the first parameter is an empty list. That
reprezentation might help with the definitional things, but as the third
axiom shows, it's not always so simple, so I don't believe this is
enough of a reason to switch away from the meta reprezentation.