Formalising Kore - domain values, simplification tactics, axiom satisfaction (#452)
* Initial generation of Kore

* Add sort axioms to theories

* Fixing Sorts symbol

* Add missing sorted quatifiers

* Kore syntax finished

* Fix proof system names

* wf tactic failure

* Fixes to wf_auto2

* Non-binder Kore wf and simplification classes

* Fix mu, nu according to the PoP paper

* Minor improvements for some equivalence theorems

* Kore fixes (in conj, disj, ceil) + equivalences

* Eq elim in Kore

* Experimental tactics for well-formedness

* Progress with original Kore, start inconsistency issue

* Imp-equiv

* Fix Kore syntax, prove equivalences, start derived rules

* Changed OPMLPattern to be dependent

Also includes temporary example.

* Test for closed pattern

Numbers need to be explicitly written in Check,
might be problematic elsewhere too.

* Investagate unsorted syntax + sorting judgement - substitution

* Fix example (unsorted approach)

* Made closed patterns more usable and renamed them

* Progress with unsorted syntax, dependently-typed semantics

* Consistency proof finished

* Proof of consistency of a forall-based axiom

* Missing file from prev commit, sort inference

* Minor fixes in sort inference

* Kore syntax - first approach

* Eliminate all admits, create general string variable signature

* Sort out Makefile, progress with syntax (sortedness)

* Refactor and extend OPML to Kore (semantics WIP)

* Dep-typed syntax of Kore

* Substitution complete

* Semantics finished

* Start working on Bool theory + model

* Kore bool theory - negation axioms

* Progress with bool theory

* Automatic proof of functional axioms

* Bool equality and non-equality

* Add Int theory to Bool

* Renamings

* Int model except rand

* Add demo, finish int model, start proving the theory cons.

* More supporting features for reasoning about ints

* Further progress with Int

* Code cleanup

* Extend Demo

* Sketches about list syntax and model

* Order-sorted signature

* Minor progress with subsorting examples

* Progress with subsorts

First I tried a version of the carrier that mirrors the subsort
relation, but it didn't seem to be very helpful and didn't solve the
problem of not being able to do inversion on Prop.

So I decided to see what would happen if we used a computable relation
(crelation) for subsorting, which did allow me to do inversion and
define the model. For this, I used the original carrier as that seems
easier to generate automatically.

I tried some axioms, and one was provable but the others are really
ugly. I'm not sure I even defined them correctly (see ? comments) but I
suspect they won't simplify because of the ltac definition. However, due
to the excessive use of dependent destructions/inductions I'm not sure
if a gallina definition would look much better.

I also tried an EqDecision definition, should be possible, but
transports don't make it easy, and I didn't want to spend more time on
it. Only needed for list inclusion.

* Alternative definition

This definition uses meta lists in the carrier. Model definition is
easier, and satisfiablity should be too, but it's worse. I don't know
why. There also seem to be no support in stdpp for fmap on propsets.

* Minimal progress with EqDec

* More progress

- finished EqDecision
  - subsort relation needs to be unique
- rewrote many things to gallina
  - dependent induction/destruction, inversion and discriminate still
    remain as ltac

None of this helped much with satisfiability.

* Closer to understanding problem

The second axiom simplifies down to app without problems, but since one
of its parameters is an evar_valuation, all the dependent inductions
don't simplify and it creates a mess.

With the non-meta list approach this didn't come up because it was
enough to know that the first parameter is an empty list. That
reprezentation might help with the definitional things, but as the third
axiom shows, it's not always so simple, so I don't believe this is
enough of a reason to switch away from the meta reprezentation.

* Third approach

Use inductives to define the separate carriers for sorts and use a
definition to map the sorts to these.

* Potential inconsistency with diamond subsorts

* Consistency of list theory

* Baz

* Comments or Baz

* Fix basic_simplify_krule, prove isList axiom

* Partial models

* Add models for maps

* Update tacs

* Added a smaller MInt example on this branch

* Fix compilation, get rid of unused classes

* Extended syntax and model, removed existing theories from _CoqProject

* Created a simple NatBool model to demonstrate dv

* Add balanced tree example

* Added more parsers and stuff

* Fixed theories broken by dv extension

* Add modeling functions to builtins (WIP, INT and BOOL is done)

* Add parsers, complete proof about bools

* Add missing file

* Add more funs

* Added map and set metatheory

* Update builtins

* Update functional solving

* Added most string functions

They may not match the hooked ones, I can't tell due to SEVERE lack of
documentation.

* Added the missing String ones I could do

* Update BIFs, simplification support

* Update solver tacs

* Tactics are smarter

* Add tree2

* Minor improvements - 14 axioms remain

* Partial builtin proofs

* App_ext empty

* Newest tree

* Proved app_ext_singleton_empty

* Fix tree example

* Improved tactics

* Structure tacs

* Tree v5 - all axioms proved

* Finally finished rem_abs_0

* Update tree.k

* Update

* Cleanup, structure

* Update _CoqProject

* Update README.md

* Fixes for compilation

* Update README.md

* Added instructions to kore README

---------

Co-authored-by: Kurucz Ádám <cphfw1@inf.elte.hu>
Co-authored-by: Dániel Horpácsi <daniel-h@elte.hu>
Co-authored-by: Kurucz Ádám <41855680+Engreyight@users.noreply.github.com>